Allow sending usermessage 'PZDmgMsg' to SourceTV (no game patch required)
I added sending usermessage `PZDmgMsg` to the SourceTV client. UserMessage is responsible for the message on the left side of the screen, for example when a hunter falls from a great height onto a victim for damage (25) or about the destruction of a canister, or about the killing of a player, etc.